+++ src/main/java/org/dcsa/standards/specifications/standards/core/v102/types/EventClassifierCode.java
@@ -1,0 +1,22 @@
+package org.dcsa.standards.specifications.standards.core.v102.types;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.dcsa.standards.specifications.generator.EnumBase;
+
+@Getter
+@Schema(
+    type = "string",
+    example = "EST",
+    description = "Classifier denoting whether an event is planned, estimated or actual",
+    maxLength = 10)
+@AllArgsConstructor
+public enum EventClassifierCode implements EnumBase {
+  PLN("Planned"),
+  EST("Estimated"),
+  ACT("Actual")
+  ;
+
+  private final String valueDescription;
+}