+++ src/main/java/org/dcsa/standards/specifications/standards/core/v102/types/HouseBillOfLadingReference.java
@@ -1,0 +1,13 @@
+package org.dcsa.standards.specifications.standards.core.v102.types;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+
+@Schema(
+    type = "string",
+    maxLength = 20,
+    example = "HBOL001",
+    description =
+"""
+Unique number allocated by the Ocean Transport Intermediary (OTI) to the House Bill of Lading.
+""")
+public class HouseBillOfLadingReference {}