+++ src/main/java/org/dcsa/standards/specifications/standards/portcall/v200/model/PortCallService.java
@@ -2,8 +2,8 @@ package org.dcsa.standards.specifications.standards.portcall.v200.model;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
-import org.dcsa.standards.specifications.standards.core.v101.model.Location;
-import org.dcsa.standards.specifications.standards.core.v101.types.UniversallyUniqueID;
+import org.dcsa.standards.specifications.standards.core.v102.model.Location;
+import org.dcsa.standards.specifications.standards.core.v102.types.UniversallyUniqueID;
 
 @Data
 @Schema(description = "Port call service information")