+++ src/main/java/org/dcsa/standards/specifications/standards/portcall/v200/model/Timestamp.java
@@ -2,8 +2,8 @@ package org.dcsa.standards.specifications.standards.portcall.v200.model;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
-import org.dcsa.standards.specifications.standards.core.v101.types.DelayReasonCode;
-import org.dcsa.standards.specifications.standards.core.v101.types.FormattedDateTime;
+import org.dcsa.standards.specifications.standards.core.v102.types.DelayReasonCode;
+import org.dcsa.standards.specifications.standards.core.v102.types.FormattedDateTime;
 
 import java.util.List;